+++ packages/react-pose/src/components/PoseElement/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { isForwardRef } from 'react-is';
+//import { isForwardRef } from 'react-is';
 import { createContext, createElement } from 'react';
 import poseFactory, {
   DomPopmotionPoser,
@@ -172,7 +172,8 @@ class PoseElement extends React.PureComponent<PoseElementInternalProps> {
     const refs: RefSetters = {};
     const { elementType } = this.props;
 
-    if (typeof elementType === 'string' || isForwardRef(elementType)) {
+    if (typeof elementType === 'string') {
+      //  || isForwardRef(elementType) TODO: Fix types for isForwardRef
       refs.ref = this.setRef;
     } else {
       refs.innerRef = this.setRef;