+++ packages/react-pose/src/components/Transition/index.tsx
@@ -1,54 +1,59 @@
 import * as React from 'react';
 import { Fragment } from 'react';
 import { Props, State } from './types';
-import {
-  handleIncomingChildren,
-  makeChildList,
-  removeFromChildren
-} from './children';
+import handleChildrenTransitions from './children';
 
 class Transition extends React.Component<Props, State> {
   static defaultProps = {
     flipMove: false,
+    enterAfterExit: false,
     preEnterPose: 'exit',
     enterPose: 'enter',
     exitPose: 'exit'
   };
 
-  static getDerivedStateFromProps = (
-    props: Props,
-    { isLeaving, removeFromTree, children }: State
-  ) => {
-    const incomingChildren = makeChildList(props.children);
-
-    return {
-      incomingChildren,
-      children: handleIncomingChildren({
-        incomingChildren,
-        displayedChildren: children,
-        isLeaving,
-        removeFromTree,
-        groupProps: props
-      })
-    };
-  };
+  static getDerivedStateFromProps = handleChildrenTransitions;
 
   state: State = {
+    children: [],
     incomingChildren: [],
-    isLeaving: new Set(),
-    removeFromTree: key => () => {
-      const { isLeaving } = this.state;
-      isLeaving.delete(key);
-      this.removeFromChildren(key);
-    }
+    finishedLeaving: {},
+    scheduleChildRemoval: key => this.removeChild(key)
   };
 
-  removeFromChildren(key: string) {
-    const { children } = this.state;
+  removeChild(key: string) {
+    const { children, finishedLeaving } = this.state;
+    const { enterAfterExit, onRest } = this.props;
+    if (!finishedLeaving.hasOwnProperty(key)) return;
 
-    this.setState({
-      children: removeFromChildren(children, key)
-    });
+    finishedLeaving[key] = true;
+
+    if (
+      !Object.keys(finishedLeaving).every(
+        leavingKey => finishedLeaving[leavingKey]
+      )
+    ) {
+      return;
+    }
+
+    const targetChildren = children.filter(
+      child => !finishedLeaving.hasOwnProperty(child.key)
+    );
+
+    const newState = enterAfterExit
+      ? {
+          finishedLeaving: {},
+          ...handleChildrenTransitions(
+            { ...this.props, enterAfterExit: false },
+            { ...this.state, children: targetChildren }
+          )
+        }
+      : {
+          finishedLeaving: {},
+          children: targetChildren
+        };
+
+    this.setState(newState, onRest);
   }
 
   shouldComponentUpdate(nextProps: Props, nextState: State) {